+++ drake/examples/kuka_iiwa_arm/test/kuka_iiwa_dynamics_test.cc
@@ -6,8 +6,8 @@
 #include "drake/examples/kuka_iiwa_arm/iiwa_simulation.h"
 #include "drake/system1/Simulation.h"
 #include "drake/system1/cascade_system.h"
+#include "drake/system1/robot_state_tap.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/systems/plants/robot_state_tap.h"
 
 using drake::CascadeSystem;
 using drake::RigidBodySystem;
