+++ drake/examples/kuka_iiwa_arm/test/kuka_iiwa_gravity_compensated_position_control_test.cc
@@ -7,8 +7,8 @@
 #include "drake/system1/Simulation.h"
 #include "drake/system1/cascade_system.h"
 #include "drake/system1/gravity_compensated_pd_position_control_system.h"
+#include "drake/system1/robot_state_tap.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/systems/plants/robot_state_tap.h"
 
 using drake::AffineSystem;
 using drake::CascadeSystem;
