+++ drake/examples/kuka_iiwa_arm/test/kuka_iiwa_pd_control_test.cc
@@ -7,8 +7,8 @@
 #include "drake/system1/Simulation.h"
 #include "drake/system1/cascade_system.h"
 #include "drake/system1/pd_control_system.h"
+#include "drake/system1/robot_state_tap.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/systems/plants/robot_state_tap.h"
 
 using drake::AffineSystem;
 using Eigen::MatrixXd;
