+++ drake/systems/plants/CMakeLists.txt
@@ -37,8 +37,7 @@ drake_install_headers(
   parser_common.h
   parser_sdf.h
   parser_urdf.h
-  rigid_body_loop.h
-  robot_state_tap.h)
+  rigid_body_loop.h)
 
 pods_install_pkg_config_file(drake-rbm
   LIBS -ldrakeRBM -ldrakeCollision -ldrakeJoints -ldrakeXMLUtil
